+++ every_election/apps/organisations/admin/organisation.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+
+class OrganisationAdmin(admin.ModelAdmin):
+    search_fields = ('official_name', 'common_name', 'official_identifier')
